file check

+++ install.sh
@@ -23,7 +23,9 @@ create_symlink () {
         INSTALL_DIR="/usr/local/bin"
     fi
 
-    ln -s "$DIR/post-merge" "$DIR/.git/hooks/post-merge"
+    if [ -f "$DIR/.git/hooks/post-merge" ]; then
+        ln -s "$DIR/post-merge" "$DIR/.git/hooks/post-merge"
+    fi
 
     # do not overwrite existing file in install directory
     if [ ! -f "$INSTALL_DIR/$1" ]; then
